move wrapped cell and context implementations into core module
