--- Components/Sources/Collection/Extensions/CollectionContext+CollectionWrappedCell.swift
@@ -1,30 +1,0 @@
-//
-//  CollectionContext+CollectionWrappedCell.swift
-//  ReactiveDataDisplayManager
-//
-//  Created by Никита Коробейников on 21.07.2023.
-//
-
-import UIKit
-import Foundation
-import ReactiveDataDisplayManager
-
-extension CollectionContext: BuilderContext {
-
-    public typealias ViewType = UICollectionViewCell
-    public typealias GeneratorType = CollectionCellGenerator
-
-    public static func gen<Item>(_ type: Item.Type, model: Item.Model) -> GeneratorType where Item: BaseItem {
-        Item.rddm.collectionGenerator(with: model)
-    }
-
-    public static func gen<Item>(_ type: Item.Type, model: Item.Model) -> GeneratorType where Item: ViewType,
-                                                                                                Item: BaseItem {
-        Item.rddm.baseGenerator(with: model, and: Item.prefferedRegistration)
-    }
-
-    public static func stack(model: StackView.Model) -> CollectionCellGenerator {
-        StackView.rddm.collectionGenerator(with: model)
-    }
-
-}
