+++ Components/Sources/Collection/Extensions/CollectionContext+StackView.swift
@@ -1,0 +1,16 @@
+//
+//  StackView+CollectionContext.swift
+//  ReactiveChat_iOS
+//
+//  Created by Никита Коробейников on 28.07.2023.
+//
+
+import ReactiveDataDisplayManager
+
+public extension CollectionContext {
+
+    static func stack(model: StackView.Model) -> CollectionCellGenerator {
+        StackView.rddm.collectionGenerator(with: model)
+    }
+
+}
