+++ Source/Collection/Generators/CollectionContext.swift
@@ -8,4 +8,18 @@
 import UIKit
 import Foundation
 
-public struct CollectionContext { }
+public struct CollectionContext: BuilderContext {
+
+    public typealias ViewType = UICollectionViewCell
+    public typealias GeneratorType = CollectionCellGenerator
+
+    public static func gen<Item>(_ type: Item.Type, model: Item.Model) -> GeneratorType where Item: BaseItem {
+        Item.rddm.collectionGenerator(with: model)
+    }
+
+    public static func gen<Item>(_ type: Item.Type, model: Item.Model) -> GeneratorType where Item: ViewType,
+                                                                                                Item: BaseItem {
+        Item.rddm.baseGenerator(with: model, and: Item.prefferedRegistration)
+    }
+
+}
