+++ Source/Table/Wrapper/TableWrappedCell.swift
@@ -6,7 +6,6 @@
 //
 
 import UIKit
-import ReactiveDataDisplayManager
 
 /// Empty table cell with `View`. Configuration is implemented within `ViewWrapper`.
 public final class TableWrappedCell<View: BaseItem>: UITableViewCell, ViewWrapper {
